Dodawanie, zmienianie, usuawnie pracownika, szczegoly pracownika (przypisane projekty, notatki, taski), przypisywanie do/usuwanie pracowniak z projektu

+++ server/src/modules/users/usersController.ts
@@ -22,7 +22,11 @@ export const getAllUsers = async (req: Request, res: Response) => {
     try {
         const users = await prisma.users.findMany({
             include: {
-                UserData: true,
+                UserData: {
+                    include: {
+                        Position: true
+                    }
+                },
                 UserRole: {
                     include: {
                         Role: true
@@ -37,13 +41,131 @@ export const getAllUsers = async (req: Request, res: Response) => {
     }
 };
 
+export const selectAllCompanyPositions = async (req: Request, res: Response) => {
+    try {
+            const positions = await prisma.companyPositions.findMany({
+        });
+        res.status(200).json(positions);
+    } catch (error) {
+        console.error("Error fetching positions:", error);
+        res.status(500).json({ error: 'Internal Server Error' });
+    }
+};
+
+export const assignUserToProject = async (req: Request, res: Response) => {
+    const { user_id } = req.params;
+    const { project_id } = req.body;
+
+    const userId = parseInt(user_id);
+    const projectId = parseInt(project_id);
+
+    // Check if user exists
+    const userExists = await prisma.users.findUnique({
+        where: { user_id: userId }
+    });
+
+    if (!userExists) {
+        return res.status(404).json({ error: 'User not found' });
+    }
+
+    // Check if project exists
+    const projectExists = await prisma.projects.findUnique({
+        where: { project_id: projectId }
+    });
+
+    if (!projectExists) {
+        return res.status(404).json({ error: 'Project not found' });
+    }
+
+    // Check if the user is already assigned to the project
+    const existingAssignment = await prisma.projectAssignments.findFirst({
+        where: {
+            user_id: parseInt(user_id),
+            project_id: parseInt(project_id)
+        }
+    });
+
+    if (existingAssignment) {
+        return res.status(400).json({ error: 'User already assigned to this project' });
+    }
+
+    try {
+        const newProjectAssignment = await prisma.projectAssignments.create({
+            data: {
+                user_id: userId,
+                project_id: projectId
+            },
+            include: {
+                User: true,
+                Project: true
+            }
+        });
+
+        res.status(201).json({ newProjectAssignment });
+    } catch (error) {
+        console.error(`Error assigning user id ${userId} to project id ${projectId}`, error);
+        res.status(500).json({ error: 'Internal Server Error' });
+    }
+};
+
+export const removeUserFromProject = async (req: Request, res: Response) => {
+    const { user_id, project_id } = req.params;
+
+    // Check if user and project exist
+    const userExists = await prisma.users.findUnique({
+        where: { user_id: parseInt(user_id) }
+    });
+
+    if (!userExists) {
+        return res.status(404).json({ error: 'User not found' });
+    }
+
+    const projectExists = await prisma.projects.findUnique({
+        where: { project_id: parseInt(project_id) }
+    });
+
+    if (!projectExists) {
+        return res.status(404).json({ error: 'Project not found' });
+    }
+
+    try {
+        // Find the project assignment to delete
+        const assignmentToDelete = await prisma.projectAssignments.findFirst({
+            where: {
+                user_id: parseInt(user_id),
+                project_id: parseInt(project_id)
+            }
+        });
+
+        if (!assignmentToDelete) {
+            return res.status(404).json({ error: 'User is not assigned to this project' });
+        }
+
+        // Delete the project assignment
+        await prisma.projectAssignments.delete({
+            where: {
+                project_assignment_id: assignmentToDelete.project_assignment_id
+            }
+        });
+
+        res.status(200).json({ message: 'User removed from project successfully' });
+    } catch (error) {
+        console.error(`Error removing user id ${user_id} from project id ${project_id}`, error);
+        res.status(500).json({ error: 'Internal Server Error' });
+    }
+};
+
 export const getUserById = async (req: Request, res: Response) => {
     const { id } = req.params;
     try {
         const user = await prisma.users.findUnique({
             where: { user_id: parseInt(id) },
             include: {
-                UserData: true,
+                UserData: {
+                    include: {
+                        Position: true
+                    }
+                },
                 UserRole: {
                     include: {
                         Role: true,
@@ -52,6 +174,11 @@ export const getUserById = async (req: Request, res: Response) => {
                 Client: true,
                 Task: true,
                 Note: true,
+                ProjectAssignment: {
+                    include: {
+                        Project: true
+                    }
+                }
             },
         });
         if (!user) {
@@ -68,7 +195,7 @@ export const createUser = async (req: Request, res: Response) => {
     const { username, email, password, userData, userRoles } = req.body;
 
     try {
-        // Sprawdzenei dostepnosci username
+        // Sprawdzenie dostępności username
         const existingUserByUsername = await prisma.users.findUnique({
             where: { username },
         });
@@ -77,7 +204,7 @@ export const createUser = async (req: Request, res: Response) => {
             return res.status(400).json({ error: 'Username already exists' });
         }
 
-        // Sprawdzenei istnienia maila
+        // Sprawdzenie istnienia emaila
         const existingUserByEmail = await prisma.users.findUnique({
             where: { email },
         });
@@ -86,7 +213,7 @@ export const createUser = async (req: Request, res: Response) => {
             return res.status(400).json({ error: 'Email already exists' });
         }
 
-        // Sprawdzenie istnienia roli
+        // Sprawdzenie istnienia roli w firmie
         const roleChecks = await Promise.all(userRoles.map(async (role: { role_id: number }) => {
             const existingRole = await prisma.roles.findUnique({
                 where: { role_id: role.role_id },
@@ -98,6 +225,20 @@ export const createUser = async (req: Request, res: Response) => {
             return res.status(400).json({ error: 'One or more roles do not exist' });
         }
 
+        // Sprawdzenie istnienia pozycji w firmie
+        const positionIds = Array.isArray(userData) ? userData.map(data => data.position_id) : [userData.position_id];
+
+        const positionChecks = await Promise.all(positionIds.map(async (position_id) => {
+            const existingPosition = await prisma.companyPositions.findUnique({
+                where: { position_id },
+            });
+            return existingPosition !== null;
+        }));
+
+        if (positionChecks.includes(false)) {
+            return res.status(400).json({ error: 'One or more positions do not exist within the company' });
+        }
+
         const hashedPassword = await bcrypt.hash(password, 10);
 
         const newUser = await prisma.users.create({
@@ -106,7 +247,7 @@ export const createUser = async (req: Request, res: Response) => {
                 email,
                 password: hashedPassword,
                 UserData: {
-                    create: userData,
+                    create: Array.isArray(userData) ? userData : [userData],
                 },
                 UserRole: {
                     create: userRoles.map((role: { role_id: number }) => ({ role_id: role.role_id })),
@@ -115,9 +256,14 @@ export const createUser = async (req: Request, res: Response) => {
             include: {
                 UserRole: {
                     include: {
-                        Role: true,
+                        Role: true
                     },
                 },
+                UserData: {
+                    include: {
+                        Position: true
+                    }
+                }
             },
         });
 
@@ -131,19 +277,81 @@ export const createUser = async (req: Request, res: Response) => {
 export const updateUser = async (req: Request, res: Response) => {
     const { id } = req.params;
     const { username, email, password, userData, userRoles } = req.body;
-    try {
-        const dataToUpdate: any = {
-            username,
-            email,
-            UserData: {
-                update: userData,
-            },
-            UserRole: {
-                deleteMany: { user_id: parseInt(id) }, // Usuń istniejące role użytkownika
-                create: userRoles.map((role: { role_id: number }) => ({ role_id: role.role_id })),
-            },
-        };
 
+    try {
+        const dataToUpdate: any = {};
+
+        // Sprawdzenie dostępności username
+        if (username) {
+            const existingUserByUsername = await prisma.users.findUnique({
+                where: { username },
+            });
+
+            if (existingUserByUsername && existingUserByUsername.user_id !== parseInt(id)) {
+                return res.status(400).json({ error: 'Username already exists' });
+            }
+
+            dataToUpdate.username = username;
+        }
+
+        // Sprawdzenie dostępności emaila
+        if (email) {
+            const existingUserByEmail = await prisma.users.findUnique({
+                where: { email },
+            });
+
+            if (existingUserByEmail && existingUserByEmail.user_id !== parseInt(id)) {
+                return res.status(400).json({ error: 'Email already exists' });
+            }
+
+            dataToUpdate.email = email;
+        }
+
+        // Sprawdzenie istnienia roli
+        if (userRoles) {
+            const roleChecks = await Promise.all(userRoles.map(async (role: { role_id: number }) => {
+                const existingRole = await prisma.roles.findUnique({
+                    where: { role_id: role.role_id },
+                });
+                return existingRole !== null;
+            }));
+
+            if (roleChecks.includes(false)) {
+                return res.status(400).json({ error: 'One or more roles do not exist' });
+            }
+
+            dataToUpdate.UserRole = {
+                deleteMany: { user_id: parseInt(id) },
+                create: userRoles.map((role: { role_id: number }) => ({ role_id: role.role_id })),
+            };
+        }
+
+        // Sprawdzenie istnienia stanowiska w firmie
+        if (userData && userData.position_id) {
+            const existingPosition = await prisma.companyPositions.findUnique({
+                where: { position_id: userData.position_id },
+            });
+
+            if (!existingPosition) {
+                return res.status(400).json({ error: 'Position does not exist' });
+            }
+
+            dataToUpdate.UserData = {
+                update: {
+                    where: { user_id: parseInt(id) },
+                    data: userData,
+                },
+            };
+        } else if (userData) {
+            dataToUpdate.UserData = {
+                update: {
+                    where: { user_id: parseInt(id) },
+                    data: userData,
+                },
+            };
+        }
+
+        // Aktualizacja hasła jesli podane
         if (password) {
             dataToUpdate.password = await bcrypt.hash(password, 10);
         }
@@ -152,7 +360,11 @@ export const updateUser = async (req: Request, res: Response) => {
             where: { user_id: parseInt(id) },
             data: dataToUpdate,
             include: {
-                UserData: true,
+                UserData: {
+                    include: {
+                        Position: true,
+                    },
+                },
                 UserRole: {
                     include: {
                         Role: true,
@@ -197,6 +409,9 @@ export const deleteUser = async (req: Request, res: Response) => {
                 where: { user_id: userId }
             }),
             prisma.notes.deleteMany({
+                where: { user_id: userId }
+            }),
+            prisma.projectAssignments.deleteMany({
                 where: { user_id: userId }
             }),
             // Usuń użytkownika
